Refactor main.rs: Extract Settings Window to UI Module
Part 1 of the main.rs refactoring plan.
- Created `docs/05-DEVELOPMENT/REFACTORING_PLAN.md` with visual analysis and roadmap.
- Created `crates/mapmap/src/ui/` module structure.
- Extracted Settings Window logic to `crates/mapmap/src/ui/settings.rs`.
- Integrated `ui::settings::show` into `main.rs`.
- Cleaned up imports and dependencies.
