+++ crates/mapmap/src/ui/mod.rs
@@ -1,0 +1,5 @@
+//! UI modules for MapFlow.
+//!
+//! This module contains the user interface components extracted from the main application loop.
+
+pub mod settings;
